update to pyrogram 0.16.0 use fstrings instead of format

+++ myself.py
@@ -15,49 +15,48 @@ import utils
 logging.basicConfig(format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
                     level=logging.ERROR)
 
-info_string = "INFO\nID: {0}\nType: {1}\nUsername: {2}\nName: {3}\nSurname: {4}\nDescription: {5}\nDate: {6}"
-
 # check telegram api key
 app = pyrogram.Client(session_name="Pyrogram",
-                      workers=2)
+                      workers=4)
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meexec", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meexec", prefixes=["/", "!", "#", "."]))
 def CmdExec(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     expression = " ".join(msg.command[1:])
 
     if expression:
         text = None
         try:
-            text = str(exec(expression,
-                            {"client": client, "msg": msg}))
-        except Exception as error:
-            text = str(error)
+            text = exec(expression,
+                        {"client": client, "msg": msg})
+        except Exception as ex:
+            text = str(ex)
 
-        if len(text) > 4096:
-            file_name = "message_too_long_{0}.txt".format(str(time.time()))
-            with open(file_name, "w") as f:
-                f.write(text)
-            msg.reply_document(document=file_name)
-            os.remove(file_name)
-        else:
-            msg.reply_text(text=text,
-                           parse_mode=None)
+        if text:
+            text = str(text)
+            if len(text) > 4096:
+                file_name = f"message_too_long_{time.time()}.txt"
+                with open(file_name, "w", encoding="utf-8") as f:
+                    f.write(text)
+                msg.reply_document(document=file_name)
+                os.remove(file_name)
+            else:
+                msg.reply_text(text=text)
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meeval", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meeval", prefixes=["/", "!", "#", "."]))
 def CmdEval(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     expression = " ".join(msg.command[1:])
 
@@ -66,28 +65,27 @@ def CmdEval(client: pyrogram.Client,
         try:
             text = str(eval(expression,
                             {"client": client, "msg": msg}))
-        except Exception as error:
-            text = str(error)
+        except Exception as ex:
+            text = str(ex)
 
         if len(text) > 4096:
-            file_name = "message_too_long_{0}.txt".format(str(time.time()))
-            with open(file_name, "w") as f:
+            file_name = f"message_too_long_{time.time()}.txt"
+            with open(file_name, "w", encoding="utf-8") as f:
                 f.write(text)
             msg.reply_document(document=file_name)
             os.remove(file_name)
         else:
             print("TEXT IS => " + text)
-            msg.reply_text(text=text,
-                           parse_mode=None)
+            msg.reply_text(text=text)
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mehelp", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mehelp", prefixes=["/", "!", "#", "."]))
 def CmdHelp(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     msg.edit_text(text="""HELP
 <code>!mehelp</code>: Sends this message.
@@ -101,13 +99,13 @@ def CmdHelp(client: pyrogram.Client,
                   parse_mode="html")
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("metodo", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("metodo", prefixes=["/", "!", "#", "."]))
 def CmdTodo(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     # remove the actual command
     msg.command.remove(msg.command[0])
@@ -117,49 +115,48 @@ def CmdTodo(client: pyrogram.Client,
         # reconstruct the path in case there are spaces (pyrogram.Filters.command uses spaces as default separator)
         message += " ".join(msg.command)
     client.send_message(chat_id="me",
-                        text="#TODO " + message,
-                        parse_mode=None)
+                        text="#TODO " + message)
     if msg.reply_to_message:
         msg.reply_to_message.forward(chat_id="me")
     msg.delete()
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mevardump", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mevardump", prefixes=["/", "!", "#", "."]))
 def CmdVardump(client: pyrogram.Client,
                msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
-    txt = ""
+    text = None
     try:
         if msg.reply_to_message:
-            txt = msg.reply_to_message
+            text = msg.reply_to_message
         else:
-            txt = msg
-        txt = utils.CensorPhone(msg)
-    except Exception as e:
-        txt = e
-    if len(str(txt)) > 4096:
-        file_name = "message_too_long_{0}.txt".format(str(time.time()))
-        with open(file_name, "w") as f:
-            f.write(str(txt))
-        msg.reply_document(document=file_name)
-        msg.delete()
-        os.remove(file_name)
-    else:
-        msg.edit_text(text="VARDUMP\n" + str(txt),
-                      parse_mode=None)
+            text = msg
+        text = str(utils.CensorPhone(msg))
+    except Exception as ex:
+        text = str(ex)
+    if text:
+        if len(text) > 4096:
+            file_name = f"message_too_long_{time.time()}.txt"
+            with open(file_name, "w", encoding="utf-8") as f:
+                f.write(text)
+            msg.reply_document(document=file_name)
+            msg.delete()
+            os.remove(file_name)
+        else:
+            msg.edit_text(text=f"VARDUMP\n{text}")
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("merawinfo", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("merawinfo", prefixes=["/", "!", "#", "."]))
 def CmdRawInfo(client: pyrogram.Client,
                msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     obj: pyrogram.User = None
     try:
@@ -170,23 +167,21 @@ def CmdRawInfo(client: pyrogram.Client,
             obj = str(client.get_chat(chat_id=msg.reply_to_message.user.id))
         else:
             obj = str(client.get_chat(chat_id=msg.chat.id))
-    except Exception as e:
-        obj = e
-    msg.edit_text(text="INFO\n" + str(obj),
-                  parse_mode=None)
+    except Exception as ex:
+        obj = ex
+    msg.edit_text(text="INFO\n" + str(obj))
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meinfo", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meinfo", prefixes=["/", "!", "#", "."]))
 def CmdInfo(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     # TODO ADD GROUPDATA FOR USERS
     obj: pyrogram.User = None
-    text = info_string
     try:
         if len(msg.command) > 1:
             obj = client.get_chat(chat_id=msg.command[1] if not utils.IsInt(
@@ -196,37 +191,30 @@ def CmdInfo(client: pyrogram.Client,
         else:
             obj = client.get_chat(chat_id=msg.chat.id)
 
-        text = text.format(obj.id,
-                           obj.type,
-                           obj.username,
-                           obj.first_name,
-                           obj.description,
-                           datetime.datetime.now().time())
-    except Exception as e:
-        text = "INFO\n" + str(e)
-    msg.edit_text(text=text,
-                  parse_mode=None)
+        text = f"INFO\nID: {obj.id}\nType: {obj.type}\nUsername: {obj.username}\nName: {obj.first_name}\nSurname: {obj.last_name if obj.last_name else ''}\nDescription: {obj.description}\nDate: {datetime.datetime.now().time()}"
+    except Exception as ex:
+        text = f"INFO\n{ex}"
+    msg.edit_text(text=text)
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meping", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("meping", prefixes=["/", "!", "#", "."]))
 def CmdPing(client: pyrogram.Client,
             msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
-    msg.edit_text(text="Pong!",
-                  parse_mode=None)
+    msg.edit_text(text=msg.text.replace("i", "o"))
 
 
-@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mereboot", prefix=["/", "!", "#", "."]))
+@app.on_message(pyrogram.Filters.user("me") & pyrogram.Filters.command("mereboot", prefixes=["/", "!", "#", "."]))
 def CmdReboot(client: pyrogram.Client,
               msg: pyrogram.Message):
-    print("\n[ UTC " + datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S") + " ] " +
-          (msg.chat.title if msg.chat.title else (msg.chat.first_name + ((" " + msg.chat.last_name) if msg.chat.last_name else ""))) +
-          " (" + (("@" + msg.chat.username + " ") if msg.chat.username else " ") +
-          "#chat{0}".format(msg.chat.id) + "): " +
+    print("\n[ UTC {0} ] ".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")) +
+          (msg.chat.title if msg.chat.title else (msg.chat.first_name + (f" {msg.chat.last_name}" if msg.chat.last_name else ""))) +
+          " (" + (f"@{msg.chat.username} " if msg.chat.username else " ") +
+          f"#chat{msg.chat.id}): " +
           msg.text)
     msg.delete()
     python = sys.executable
@@ -234,8 +222,10 @@ def CmdReboot(client: pyrogram.Client,
 
 
 app.start()
+app.set_parse_mode(parse_mode=None)
 user = app.get_me()
 print(user)
 app.send_message(chat_id=user.id,
                  text="Bot started!\nUTC {0}".format(datetime.datetime.utcnow().strftime("%Y/%m/%d %H:%M:%S")))
 app.idle()
+app.stop()
